Preparing for "plot anything" histograms and scatter plots on the fly
Former-commit-id: ea9ef627e4efb80e69b8d93545edeee191ecb13c

+++ ced/src/main/java/cnuphys/ced/cedview/bst/BSTzHitDrawer.java
@@ -130,7 +130,7 @@ public class BSTzHitDrawer implements IDrawable {
 							fract = ((double) hitCount) / maxHit;
 						}
 						else {
-							fract = Math.log((double) (hitCount + 1.))
+							fract = Math.log(hitCount + 1.)
 									/ Math.log(maxHit + 1.);
 						}
