+++ ced/src/main/java/cnuphys/ced/item/PCALHexSectorItem.java
@@ -175,7 +175,7 @@ public class PCALHexSectorItem extends HexSectorItem {
 							fract = ((double) hit) / maxHit;
 						}
 						else {
-							fract = Math.log((double) (hit + 1.))
+							fract = Math.log(hit + 1.)
 									/ Math.log(maxHit + 1.);
 						}
